Adds storage for credentials
The first time that a user calls
`GoogleAuthentication.authenticate(...)` they will be made to
authenticate with Google.  Their credentials will be stored in a local
file so that they won't need to reauth each run.

I also removed code that is no longer needed dealing with reading the
PKI file.