Documents `UASCalendar` class and adds some tests
I added DocStrings to the `UASCalendar` class to explain the methods.  I
also started writing tests to cover that class which required me to
write a mock google interface.  Finally, since these files shouldn't be
called by themselves but rather included in another python file, I've
removed the code that ran when this file is called from the CLI.